# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Architectural Portfolio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  projectCount?: number | null;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
}

export interface ProjectImage {
  id: number;
  projectId: number;
  imageUrl: string;
  isHero: boolean;
  sortOrder: number;
}

export interface ProjectImageInput {
  /** @minLength 1 */
  imageUrl: string;
  isHero?: boolean;
  sortOrder?: number;
}

export interface ProjectImageUpdate {
  imageUrl?: string;
  isHero?: boolean;
  sortOrder?: number;
}

export interface ProjectSummary {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  client?: string | null;
  /** @nullable */
  sector?: string | null;
  status: string;
  published: boolean;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  heroImage?: string | null;
  /** @nullable */
  year?: string | null;
}

export interface Project {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  client?: string | null;
  /** @nullable */
  sector?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  scope?: string | null;
  status: string;
  published: boolean;
  /** @nullable */
  longDescription?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  heroImage?: string | null;
  /** @nullable */
  year?: string | null;
  images?: ProjectImage[];
}

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  slug: string;
  location?: string;
  client?: string;
  sector?: string;
  size?: string;
  scope?: string;
  status: string;
  longDescription?: string;
  /** @nullable */
  categoryId?: number | null;
  year?: string;
  published?: boolean;
}

export interface PublishToggle {
  published: boolean;
}

export interface AdminCredentials {
  /** @minLength 1 */
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
}

export type ListProjectsParams = {
published?: boolean;
category_id?: number;
};