# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Architectural Portfolio API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: categories
    description: Project categories
  - name: projects
    description: Portfolio projects
  - name: images
    description: Project images
  - name: admin
    description: Admin authentication
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: published
          in: query
          schema:
            type: boolean
        - name: category_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/featured:
    get:
      operationId: listFeaturedProjects
      tags: [projects]
      summary: List featured published projects for homepage
      responses:
        "200":
          description: Featured projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"

  /projects/{slug}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found

  /projects/{id}/update:
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}/publish:
    patch:
      operationId: toggleProjectPublish
      tags: [projects]
      summary: Toggle project published status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishToggle"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}/delete:
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/{id}/images:
    get:
      operationId: listProjectImages
      tags: [images]
      summary: List images for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectImage"
    post:
      operationId: addProjectImage
      tags: [images]
      summary: Add an image to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectImageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectImage"

  /projects/{id}/images/{imageId}:
    put:
      operationId: updateProjectImage
      tags: [images]
      summary: Update a project image (reorder, set hero)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectImageUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectImage"
    delete:
      operationId: deleteProjectImage
      tags: [images]
      summary: Delete a project image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Unauthorized

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"

  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Check admin session
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        projectCount:
          type: ["integer", "null"]

    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1

    ProjectImage:
      type: object
      required: [id, projectId, imageUrl, isHero, sortOrder]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        imageUrl:
          type: string
        isHero:
          type: boolean
        sortOrder:
          type: integer

    ProjectImageInput:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
          minLength: 1
        isHero:
          type: boolean
        sortOrder:
          type: integer

    ProjectImageUpdate:
      type: object
      properties:
        imageUrl:
          type: string
        isHero:
          type: boolean
        sortOrder:
          type: integer

    ProjectSummary:
      type: object
      required: [id, title, slug, status, published]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        location:
          type: ["string", "null"]
        client:
          type: ["string", "null"]
        sector:
          type: ["string", "null"]
        status:
          type: string
        published:
          type: boolean
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        heroImage:
          type: ["string", "null"]
        year:
          type: ["string", "null"]

    Project:
      type: object
      required: [id, title, slug, status, published]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        location:
          type: ["string", "null"]
        client:
          type: ["string", "null"]
        sector:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        scope:
          type: ["string", "null"]
        status:
          type: string
        published:
          type: boolean
        longDescription:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        heroImage:
          type: ["string", "null"]
        year:
          type: ["string", "null"]
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProjectImage"

    ProjectInput:
      type: object
      required: [title, slug, status]
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        location:
          type: string
        client:
          type: string
        sector:
          type: string
        size:
          type: string
        scope:
          type: string
        status:
          type: string
        longDescription:
          type: string
        categoryId:
          type: ["integer", "null"]
        year:
          type: string
        published:
          type: boolean

    PublishToggle:
      type: object
      required: [published]
      properties:
        published:
          type: boolean

    AdminCredentials:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 1

    AdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean