# lib/api-zod/src/generated/types/adminCredentials.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Architectural Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminCredentials {
  /** @minLength 1 */
  password: string;
}