# lib/api-zod/src/generated/types/projectInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Architectural Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  slug: string;
  location?: string;
  client?: string;
  sector?: string;
  size?: string;
  scope?: string;
  status: string;
  longDescription?: string;
  /** @nullable */
  categoryId?: number | null;
  year?: string;
  published?: boolean;
}